+++ resources/[dg]/dg-api/typescript/src/server/routes/index.ts
@@ -1,18 +1,20 @@
-import { registerRoute } from 'sv_routes';
 import './info';
 import './token';
 import './financials';
 import './business';
 import './admin';
 import './info/character';
+import { server } from 'sv_router';
 
-registerRoute('GET', '/', (_, res) => {
-  res(200, {
+server.get('/', (_, res) => {
+  res.code(200).send({
     host: ExecuteCommand('sv_projectName'),
   });
 });
 
-registerRoute('POST', '/post-test', (req, res) => {
-  console.log('POST PONG', req.body);
-  res(200, {});
+server.post<{ Body: { model: string; owner: number } }>('/vehicles/give', async (req, res) => {
+  await global.exports['dg-vehicles'].giveNewVehicle(req.body.model, Number(req.body.owner));
+  res.code(200).send({
+    result: true,
+  });
 });