+++ resources/[dg]/dg-api/typescript/src/server/server.ts
@@ -1,3 +1,2 @@
 import './sv_router';
-import './sv_routes';
 import './routes';